Solution to Client/Server Cache Desync (Cache No longer Exists technically... we just use File IO as little as necessary)
Server can now be paused as an added security measure. This allows the Admins to edit the YAML files without risk of the Server overriding them. As a side effect, Update files will now need to be deleted before a new one can be created, Scorekeeper will block until it is deleted by WriteServer, this is another synchronization measure. It becomes difficult to handle when we are trying to mimic Network IO (which has protocols) using File IO.

Also some more documentation, cleanup and minor Conversion to Ruby idiomatic programming.

I do have a dissent in WriteServer and I would like to note how this system is not scalable nor able to adapt to changes in setup should that be necessary.